Tests/LibWeb/Text/expected/css/font-face-with-prelude-is-invalid.txt

`@font-face foo {}` should be invalid: PASS
